fix: refactor file structure and update docs
