+++ packages/popover/package.json
@@ -11,24 +11,17 @@
   },
   "author": "Priceline",
   "license": "MIT",
-  "jest": {
-    "coverageReporters": [
-      "lcov",
-      "html"
-    ],
-    "testMatch": [
-      "<rootDir>/test/**/*.js"
-    ]
-  },
   "dependencies": {
     "pcln-design-system": "^2.0.4",
     "react-popper": "^1.3.3"
   },
   "peerDependencies": {
-    "pcln-design-system": ">=1.x || >=2.x",
+    "pcln-design-system": ">=1.0.0 || >=2.0.0",
     "styled-components": ">=3.x || >= 4.x",
-    "react": ">=16.3.0",
-    "react-dom": ">=16.3.0"
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "prop-types": "^15.5.10",
+    "styled-system": "^3.1.6"
   },
   "devDependencies": {
     "@babel/plugin-transform-runtime": "^7.4.4",
