--- packages/popover/src/arrow/index.js
@@ -1,2 +1,0 @@
-import PopoverArrow from './Arrow'
-export default PopoverArrow
