--- packages/popover/src/overlay/index.js
@@ -1,2 +1,0 @@
-import Overlay from './Overlay'
-export default Overlay
