--- packages/popover/src/popover/index.js
@@ -1,2 +1,0 @@
-import Popover from './Popover'
-export default Popover
